+++ cinn/runtime/cpu/host_intrinsics.h
@@ -55,7 +55,7 @@ inline int FN_INT32(clz)(int x);
 
 #define FN_INT64(func) cinn_host_##func##_uint64
 
-inline int FN_INT64(clz)(long long x);
+inline int64_t FN_INT64(clz)(int64_t x);
 
 #undef FN_INT64
 }
